Move TransitionTarget to its own file
