+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/navigation/TransitionTarget.kt
@@ -1,0 +1,10 @@
+package com.stripe.android.paymentsheet.navigation
+
+internal sealed interface TransitionTarget {
+
+    object SelectSavedPaymentMethods : TransitionTarget
+
+    object AddAnotherPaymentMethod : TransitionTarget
+
+    object AddFirstPaymentMethod : TransitionTarget
+}
